Additional ignored channels #109

+++ src/events/newMessage.js
@@ -161,7 +161,7 @@ module.exports = {
                     }
                 }
             }
-            if (message.content.includes('discord.gg/') || message.content.includes('discordapp.com/invite/')) {
+            if (message.content.includes('discord.gg/') || message.content.includes('discordapp.com/invite/') && message.channel.parentID != "360838298677149720") {
                 if (permits[message.author.id] && permits[message.author.id].channel == message.channel.id) {
                     if (permits[message.author.id].until < new Date().getTime()) return message.delete();
                 } else {
@@ -244,7 +244,7 @@ module.exports = {
             })    
         }
 
-        if(!hasRole(message.member, 'Staff') && !hasRole(message.member, "Moderator") && !hasRole(message.member, "Admin") && !hasRole(message.member, "Developer") && !hasRole(message.member, "Community Helper")) {      
+        if(!hasRole(message.member, 'Staff') && !hasRole(message.member, "Moderator") && !hasRole(message.member, "Admin") && !hasRole(message.member, "Developer") && !hasRole(message.member, "Community Helper") && message.channel.parentID != "360838298677149720") {      
             if (!AllowedLinksSet.has(message.channel.id)) {
                 args.forEach(word => {
                     if(new RegExp("(https?:\/\/(?:www\.|(?!www))[a-zA-Z0-9][a-zA-Z0-9-]+[a-zA-Z0-9]\.[^\s]{2,}|www\.[a-zA-Z0-9][a-zA-Z0-9-]+[a-zA-Z0-9]\.[^\s]{2,}|https?:\/\/(?:www\.|(?!www))[a-zA-Z0-9]\.[^\s]{2,}|www\.[a-zA-Z0-9]\.[^\s]{2,})").test(word)) {
